Refactor class hierarchy
